added the front end and the chakra framwork

+++ backend/controllers/product.controller.js
@@ -58,11 +58,18 @@ export const updateProducts = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "invalid product ID" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "product deleted" });
   } catch (error) {
     console.log("error in deleting product");
-    res.status(404).json({ success: false, message: "product ID not found" });
+    res.status(500).json({ success: false, message: "server error" });
   }
 };
